数据格式、引入编码

+++ backend/ruoyi-system/src/main/java/com/ruoyi/system/domain/IcesMaterialStore.java
@@ -38,11 +38,11 @@ public class IcesMaterialStore extends BaseEntity {
     /**
      * 货位数量
      */
-    private Long msSpace;
+    private Float msSpace;
     /**
      * 空闲货位数量
      */
-    private Long msFree;
+    private Float msFree;
     /**
      * 已删除
      */
